El punto dos esta acabado (menos el extra y lo del menu de la decision que esta comentado en el propio codigo)

+++ Metodes/AgendaProves/Program.cs
@@ -1,58 +1,90 @@
-﻿RecollidaDades("");
-static void RecollidaDades(string dades)
+﻿Console.Write("Que vols fer: ");
+int aux = Convert.ToInt32(Console.ReadLine());
+switch (aux)
+{
+    //Introducir dades (1)
+    case 1:
+        RecollidaDades("");
+        break;
+    //Recuperar usuari (2)
+    case 2:
+        Console.Write("Indica el nom d'usuari: ");
+        string usuari = Convert.ToString(Console.ReadLine());
+        usuari = CorregirNoms(usuari);
+        if (TrobarUsuari(usuari) == "")
+            Console.WriteLine("No s'ha trobat cap usuari amb aquest usuari, vols buscar un altre? "); //Yo aqui volveria al menu.
+        //resposta = convert.ToString(Console.ReadLine());
+        //if resposta es no se que vuelve a pedir el swtich
+        //si no da igual y vuelve al menu
+        else Console.WriteLine(TrobarUsuari(usuari));
+        break;
+    default:
+        Console.WriteLine("hola");
+        break;
+}
+static void RecollidaDades(string dades) //Angel: Quiero que cada vez
 {
     string nom = "", cognom1 = "", cognom2 = "", dni = "", telefon = "", data = "", correu = "";
     char resposta = ' ';
+    BorrarConsola();
     Console.WriteLine("Indica el teu nom: ");
     nom = Convert.ToString(Console.ReadLine());
     nom = CorregirNoms(nom);
     dades += nom + " ";
-    Console.WriteLine("Tens mes d'un cognom? (S/N): ");
+    BorrarConsola();
+    Console.Write("Tens mes d'un cognom? (S/N): ");
     resposta = Convert.ToChar(Console.ReadLine());
     resposta = char.ToUpper(resposta);
     if (resposta == 'S')
     {
-        Console.WriteLine("Indica el primer cognom: ");
+        BorrarConsola();
+        Console.Write("Indica el primer cognom: ");
         cognom1 = Convert.ToString(Console.ReadLine());
         cognom1 = CorregirNoms(cognom1);
-        Console.WriteLine("Indica el primer cognom: ");
+        BorrarConsola();
+        Console.Write("Indica el segon cognom: ");
         cognom2 = Convert.ToString(Console.ReadLine());
         cognom2 = CorregirNoms(cognom2);
         dades += cognom1 + " " + cognom2 + " ";
     }
     else
     {
-        Console.WriteLine("Indica el cognom: ");
+        BorrarConsola();
+        Console.Write("Indica el cognom: ");
         cognom1 = Convert.ToString(Console.ReadLine());
         cognom1 = CorregirNoms(cognom1);
         dades += cognom1 + " ";
     }
-    Console.WriteLine("Indica el teu dni: ");
+    BorrarConsola();
+    Console.Write("Indica el teu dni: ");
     dni = Convert.ToString(Console.ReadLine());
     dades += dni + " ";
     while (!ValidacioDNI(dni))
     {
-        Console.WriteLine("DNI erroni, introdueix un DNI valid: ");
+        Console.Write("DNI erroni, introdueix un DNI valid: ");
         dni = Convert.ToString(Console.ReadLine());
     }
-    Console.WriteLine("Indica el teu telefon: ");
+    BorrarConsola();
+    Console.Write("Indica el teu telefon: ");
     telefon = Convert.ToString(Console.ReadLine());
     while (!VerificarNumeros(telefon))
     {
-        Console.WriteLine("DNI erroni, introdueix un DNI valid: ");
+        Console.Write("Telefon erroni, introdueix un telefon valid: ");
         telefon = Convert.ToString(Console.ReadLine());
     }
     dades += telefon + " ";
-    Console.WriteLine("Indica el teu data de naixement: ");
+    BorrarConsola();
+    Console.Write("Indica el teu data de naixement: ");
     data = Convert.ToString(Console.ReadLine());
     dades += data + " ";
-    Console.WriteLine("Indica el teu correu electronic: ");
+    BorrarConsola();
+    Console.Write("Indica el teu correu electronic: ");
     correu = Convert.ToString(Console.ReadLine());
     data += correu;
     dades = DadesAddCSV(ref dades);
     AgendaWriter(dades);
 }
-static string DadesAddCSV(ref string dades)
+static string DadesAddCSV(ref string dades) //Me pone un ; doble al final, puede que haya un espacio final
 {
     string caracter = ";";
     for (int i = 0; i < dades.Length; i++)
@@ -118,12 +150,43 @@ static bool VerificarNumeros(string telefon)
 }
 static bool VerificacionData(string data)
 {
-    Console.WriteLine("hola");
+    Console.Write("hola");
     return true;
 }
 static void AgendaWriter(string linea)
 {
-    StreamWriter agendaW = new StreamWriter("agenda.txt");
+    StreamWriter agendaW;
+    if (File.Exists("agenda.txt"))
+    {
+        agendaW = new StreamWriter("agenda.txt", true);
+    }
+    else
+    {
+        agendaW = new StreamWriter("agenda.txt");
+    }
     agendaW.WriteLine(linea);
     agendaW.Close();
+}
+static string TrobarUsuari(string usuari)
+{
+    StreamReader agendaR = new StreamReader("agenda.txt");
+    bool trobat = false;
+    string linea = agendaR.ReadLine(), aux = "", nom = "";
+    while (!trobat && !agendaR.EndOfStream)
+    {
+        linea = agendaR.ReadLine();
+        aux = linea;
+        nom = linea.Substring(0, linea.IndexOf(';'));
+        if (usuari == nom)
+            trobat = true;
+    }
+    if (!trobat)
+        aux = "";
+    agendaR.Close();
+    return aux;
+}
+static void BorrarConsola() 
+{
+    Console.Clear();
+    return;
 }
